+++ CHANGELOG.md
@@ -49,6 +49,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 - Pointer methods for `SEL`.
 
+- Implemented `PartialEq` for `NSObject`.
+
 - Implemented `Default` for `NSString` and `NSMutableString`.
 
 ### Fixed